Reduce memory allocations when generating the merkle root
When generating the merkle root, the previous code calls `malloc`
once for every transaction, which is not ideal for performance.

The new code allocates a 32 KiB buffer once and reuses it for all
transactions. If a TX is bigger than the current buffer size, the
buffer is reallocated with double the current buffer size.